sprint-3/step-1 feat: работает модалка, но надо починить не модалку
