+++ src/pages/MainPage/MainPage.jsx
@@ -1,22 +1,19 @@
 import React from 'react'
-import styles from './MainPage.module.css'
-import Modal from '../../components/Modal/Modal'
-import { MODAL_INGREDIENT_TITLE } from '../../utils/constants'
-import IngredientDetails from '../../components/IngredientDetails/IngredientDetails'
-import OrderDetails from '../../components/OrderDetails/OrderDetails'
-import AppHeader from '../../components/AppHeader/AppHeader'
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
+import { useDispatch, useSelector } from 'react-redux'
+import styles from './MainPage.module.css'
+import Modal from '../../components/Modal/Modal'
+import OrderDetails from '../../components/OrderDetails/OrderDetails'
+import AppHeader from '../../components/AppHeader/AppHeader'
 import BurgerIngredients from '../../components/BurgerIngredients/BurgerIngredients'
 import BurgerConstructor from '../../components/BurgerConstructor/BurgerConstructor'
-import { useDispatch, useSelector } from 'react-redux'
-import { CLEAR_DETAIL_INGREDIENT, SET_DETAIL_INGREDIENT } from '../../services/actions/detailIngredient'
+import { SET_DETAIL_INGREDIENT } from '../../services/actions/detailIngredient'
 import { CLOSE_ORDER_MODAL, postOrder } from '../../services/actions/order'
 import { getComponents } from '../../services/actions/receivedComponents'
 import { filterBun } from '../../components/App/App'
 
 const MainPage = () => {
-  const [isIngredientModalVisible, setIsIngredientModalVisible] = React.useState(false)
   const [choosenBun, setChoosenBun] = React.useState({})
 
   // Вытащим из хранилища данные о элементах с сервера и ошибках
@@ -30,7 +27,6 @@ const MainPage = () => {
 
   const handleOpenIngredientModal = (currentIngredient) => {
     dispatch({ type: SET_DETAIL_INGREDIENT, ingredient: currentIngredient })
-    setIsIngredientModalVisible(true)
   }
 
   const handleOpenOrderModal = (req) => {
@@ -38,10 +34,7 @@ const MainPage = () => {
     dispatch(postOrder(req))
   }
 
-  const handleCloseIngredientModal = () => {
-    setIsIngredientModalVisible(false)
-    dispatch({ type: CLEAR_DETAIL_INGREDIENT })
-  }
+
 
   const handleCloseOrderModal = () => {
     dispatch({ type: CLOSE_ORDER_MODAL })
@@ -61,12 +54,7 @@ const MainPage = () => {
 
   return (
     <div className={styles.application} id="app">
-      {isIngredientModalVisible && <Modal
-        closePopup={handleCloseIngredientModal}
-        title={MODAL_INGREDIENT_TITLE}
-      >
-        <IngredientDetails />
-      </Modal>}
+
       {isOrderModalVisible &&  <Modal
         closePopup={handleCloseOrderModal}
       >
